+++ contracts/modules/SnapshotSchedulerModule.sol
@@ -7,21 +7,27 @@ import {SnapshotBase} from "../library/SnapshotBase.sol";
 import {ISnapshotScheduler} from "../interface/ISnapshotScheduler.sol";
 abstract contract SnapshotSchedulerModule is SnapshotBase, ISnapshotScheduler {
     /*//////////////////////////////////////////////////////////////
+                               MODIFIERS
+    //////////////////////////////////////////////////////////////*/
+    modifier onlySnapshotManager() {
+        _authorizeSnapshot();
+        _;
+    }
+
+    /*//////////////////////////////////////////////////////////////
                             PUBLIC/EXTERNAL FUNCTIONS
     //////////////////////////////////////////////////////////////*/
     /**
     * @inheritdoc ISnapshotScheduler
     */
-    function poke() public virtual {
-        _authorizeSnapshot();
+    function poke() public virtual onlySnapshotManager {
         SnapshotBase._setCurrentSnapshot();
     }
 
     /**
     * @inheritdoc ISnapshotScheduler
     */
-    function scheduleSnapshot(uint256 time) public virtual {
-         _authorizeSnapshot();
+    function scheduleSnapshot(uint256 time) public virtual onlySnapshotManager {
         SnapshotBase._scheduleSnapshot(time);
     }
 
@@ -30,8 +36,7 @@ abstract contract SnapshotSchedulerModule is SnapshotBase, ISnapshotScheduler {
     */
     function scheduleSnapshotNotOptimized(
         uint256 time
-    ) public virtual {
-        _authorizeSnapshot();
+    ) public virtual onlySnapshotManager {
         SnapshotBase._scheduleSnapshotNotOptimized(time);
     }
 
@@ -41,8 +46,7 @@ abstract contract SnapshotSchedulerModule is SnapshotBase, ISnapshotScheduler {
     function rescheduleSnapshot(
         uint256 oldTime,
         uint256 newTime
-    ) public virtual {
-        _authorizeSnapshot();
+    ) public virtual onlySnapshotManager {
         SnapshotBase._rescheduleSnapshot(oldTime, newTime);
     }
 
@@ -51,8 +55,7 @@ abstract contract SnapshotSchedulerModule is SnapshotBase, ISnapshotScheduler {
     */
     function unscheduleLastSnapshot(
         uint256 time
-    ) public virtual {
-        _authorizeSnapshot();
+    ) public virtual onlySnapshotManager {
         SnapshotBase._unscheduleLastSnapshot(time);
     }
 
@@ -61,8 +64,7 @@ abstract contract SnapshotSchedulerModule is SnapshotBase, ISnapshotScheduler {
     */
     function unscheduleSnapshotNotOptimized(
         uint256 time
-    ) public virtual  {
-        _authorizeSnapshot();
+    ) public virtual onlySnapshotManager {
         SnapshotBase._unscheduleSnapshotNotOptimized(time);
     }